the bug is gone because of reasons unknown to me
the only change in this commit that made the bug go away was separating out search_folder_recurs and search_for_pcks into their own variables. I don't know why that had any bearing on anything, but the bug is gone now. Also seen in this commit is that all of the io stuff is handled statically now, and the folder checks have been moved to their own function for readability.