ui is now properly partitioned
Rather than a single ui_elements array which proved to be unwieldy in the editor and hard to expand, there are now three variables, for under the health bar, to the side, and the bottom of the screen, which can be used as a developer desires. As well, there is now a _connect_hud_elements function, which will handle setup per fighter, while the _initialize_training_mode_elements function will have default behavior for all fighters.