a few more state effects and flags
new flags:
SUPER - the hitbox is part of a "Super" attack. Intended to be used to denote a flashy win
NO_KO - the damage suffered by the attack cannot reduce the attacked's health to 0. Intended to be used for multi-hitting attack animations

new effects:
COLLAPSE - this attack, upon hit, places the attacked in a slow collapsing state. See Guilty Gear Strive Boss Nagoriyuki's Kamuriyuki attack for an idea
BOUNCE - the attacked, upon being hit by the attack, will bounce off of the edge of the camera (or the wall of the stage, whichever comes first)